+++ src/error.rs
@@ -44,9 +44,6 @@ impl convert::From<io::Error> for Error {
 #[cfg(not(target_os = "wasi"))]
 impl convert::From<nix::Error> for Error {
     fn from(e: nix::Error) -> Error {
-        match e {
-            nix::Error::Sys(errno) => Error::Io(errno.into()),
-            other => Error::Unexpected(format!("{:?}", other)), // should just be dealing with errno case
-        }
+        Error::Io(e.into())
     }
 }